+++ lib/src/element/sub/text/input_formatters/money_input_formatter.dart
@@ -1,9 +1,12 @@
 import 'package:flutter/services.dart';
+import 'package:simple_widget_markup/src/element/sub/text/input_formatters/util_input_formatter.dart';
 
 /// (en) A formatter for entering monetary amounts
 /// that automatically inserts commas to separate thousands.
+/// The decimal_input_formatter and money_input_formatter must be applied first.
 ///
 /// (ja) 三桁区切りでカンマが自動入力される、金額の入力のためのフォーマッター。
+/// decimal_input_formatterやmoney_input_formatterは先に適用する必要があります。
 class MoneyInputFormatter extends TextInputFormatter {
   @override
   TextEditingValue formatEditUpdate(
@@ -19,6 +22,13 @@ class MoneyInputFormatter extends TextInputFormatter {
 
     // 変更後のテキストからカンマを除去
     String newText = newValue.text.replaceAll(',', '');
+
+    // マイナス記号があれば一時的に削除。
+    bool minusFlag = false;
+    if (newText.startsWith("-")) {
+      newText = newText.replaceAll('-', '');
+      minusFlag = true;
+    }
 
     String newFormattedText = "";
 
@@ -39,14 +49,26 @@ class MoneyInputFormatter extends TextInputFormatter {
       newFormattedText = _commaFormat(newText);
     }
 
+    // - 記号で始まる場合の例外処理。
+    if (minusFlag) {
+      newFormattedText = "-$newFormattedText";
+    }
+
     // カーソル位置をカンマの増減に応じて変更。
     String preIndexBeforeText =
-        _getIndexBeforeString(oldValue.text, selectionIndex);
+        UtilInputFormatter.getIndexBeforeString(oldValue.text, selectionIndex);
     final int preRangeCommaNum = preIndexBeforeText.split(",").length - 1;
-    String newIndexBeforeText =
-        _getIndexBeforeString(newFormattedText, selectionIndex);
+    String newIndexBeforeText = UtilInputFormatter.getIndexBeforeString(
+        newFormattedText, selectionIndex);
     final int newRangeCommaNum = newIndexBeforeText.split(",").length - 1;
-    selectionIndex -= (preRangeCommaNum - newRangeCommaNum);
+    if (oldValue.text.length < newValue.text.length) {
+      final int shift = preRangeCommaNum - newRangeCommaNum;
+      if (shift < 0) {
+        selectionIndex -= (preRangeCommaNum - newRangeCommaNum);
+      }
+    } else {
+      selectionIndex -= (preRangeCommaNum - newRangeCommaNum);
+    }
 
     // カーソル位置を新しいテキストの範囲内に制限
     if (selectionIndex < 0) {
@@ -60,19 +82,6 @@ class MoneyInputFormatter extends TextInputFormatter {
       text: newFormattedText,
       selection: TextSelection.collapsed(offset: selectionIndex),
     );
-  }
-
-  /// 指定されたインデックスまでのテキストを取得する関数。
-  String _getIndexBeforeString(String t, int index) {
-    final List<String> sList = t.split('').toList();
-    final buf = StringBuffer();
-    for (int i = 0; i < sList.length; i++) {
-      buf.write(sList[i]);
-      if (i == index) {
-        break;
-      }
-    }
-    return buf.toString();
   }
 
   /// 三桁区切りでカンマを挿入する関数。