+++ lib/src/element/sub/text/input_formatters/util_input_formatter.dart
@@ -1,0 +1,14 @@
+/// (en) A utility for input formatter.
+///
+/// (ja) インプットフォーマッター用のユーティリティクラス。
+class UtilInputFormatter {
+  /// 指定されたインデックスまで（インデックスを含まない）のテキストを取得する関数。
+  static String getIndexBeforeString(String t, int index) {
+    // indexが範囲外の場合
+    if (index > t.length) {
+      return t;
+    } else {
+      return t.substring(0, index);
+    }
+  }
+}